Refactor statistics
