+++ frontend/src/app/pages/exercises/exercise/shared/core/statistics/statistics-entry.ts
@@ -1,0 +1,17 @@
+import type { UUID } from 'digital-fuesim-manv-shared';
+import type { AreaStatistics } from './area-statistics';
+
+export interface StatisticsEntry {
+    readonly id: UUID;
+
+    /**
+     * The exercise time in ms when these statistics were created.
+     */
+    readonly exerciseTime: number;
+
+    readonly viewports: {
+        readonly [viewportId: string]: AreaStatistics;
+    };
+
+    readonly exercise: AreaStatistics;
+}
