+++ frontend/src/app/pages/exercises/exercise/shared/exercise-statistics/area-statistics.service.ts
@@ -1,10 +1,10 @@
 import { Injectable } from '@angular/core';
-import { Store } from '@ngrx/store';
 import type { UUID } from 'digital-fuesim-manv-shared';
-import type { AreaStatistics } from 'digital-fuesim-manv-shared/dist/models/utils/area-statistics';
 import type { Observable } from 'rxjs';
-import { BehaviorSubject, map, switchMap } from 'rxjs';
-import type { AppState } from 'src/app/state/app.state';
+import { BehaviorSubject, combineLatest, map, ReplaySubject } from 'rxjs';
+import type { AreaStatistics } from '../core/statistics/area-statistics';
+import type { StatisticsEntry } from '../core/statistics/statistics-entry';
+import { StatisticsService } from '../core/statistics/statistics.service';
 import { decimateStatistics } from './shared/decimate-statistics';
 
 @Injectable({
@@ -18,27 +18,30 @@ export class AreaStatisticsService {
     private areaId: UUID | null = null;
     public readonly areaId$ = new BehaviorSubject<UUID | null>(this.areaId);
 
+    private readonly statistics$ = new ReplaySubject<
+        readonly StatisticsEntry[]
+    >(1);
+
     /**
      * Emits the statistics for the current area and the accompanying correct exerciseTime
      */
     public readonly areaStatistics$: Observable<AreaStatisticsEntry[]> =
-        this.areaId$.pipe(
-            switchMap((areaId) =>
-                this.store.select((state) =>
-                    state.exercise.statistics
-                        .map((statisticEntry) => ({
-                            value:
-                                areaId === null
-                                    ? statisticEntry.exercise
-                                    : // If the viewport didn't exist yet
-                                      statisticEntry.viewports[areaId],
-                            exerciseTime: statisticEntry.exerciseTime,
-                        }))
-                        .filter((entry) => entry.value !== undefined)
-                )
+        combineLatest([this.statistics$, this.areaId$]).pipe(
+            map(([statistics, areaId]) =>
+                statistics
+                    .map((statisticEntry) => ({
+                        value:
+                            areaId === null
+                                ? statisticEntry.exercise
+                                : // If the viewport didn't exist yet
+                                  statisticEntry.viewports[areaId],
+                        exerciseTime: statisticEntry.exerciseTime,
+                    }))
+                    .filter((entry) => entry.value !== undefined)
             )
         );
 
+    // TODO: This should not be necessary anymore...
     /**
      * The {@link areaStatistics$} decimated to reduce the amount of data per emit
      */
@@ -46,12 +49,17 @@ export class AreaStatisticsService {
         map((statistics) => decimateStatistics(statistics))
     );
 
+    constructor(private readonly statisticsService: StatisticsService) {}
+
     public setAreaId(areaId: UUID | null) {
         this.areaId = areaId;
         this.areaId$.next(areaId);
     }
 
-    constructor(private readonly store: Store<AppState>) {}
+    public async updateStatistics() {
+        const statistics = await this.statisticsService.generateStatistics();
+        this.statistics$.next(statistics);
+    }
 }
 
 export interface AreaStatisticsEntry {
